requirements.txt
Flask==2.1.0
Werkzeug==2.0.3
firebase-admin==6.0.0
vertexai==1.71.1
gunicorn==20.1.0
google-cloud-secret-manager==2.16.0